requirements.txt
Django~=5.1
djangorestframework~=3.15.2
django-filter~=24.3
djangorestframework-api-key~=3.0.0
python-dotenv~=1.0.1
drf-yasg~=1.21.7
djoser==2.2.3